refactor: split HandleEvent into mouse/keyboard handlers and files
Break the ~490-line HandleEvent into a thin dispatcher that routes to
handleMouse and handleKey. Extract the single-cursor key handlers
(execEnter/execBackspace/execDelete/execRune) to mirror the existing
multiExec* methods. Move the mouse and keyboard handling into
editor_widget_mouse.go and editor_widget_keyboard.go, matching the
per-concern file convention. Pure mechanical move, no behavior change.
