+++ internal/ui/editor_widget.go
@@ -3,7 +3,6 @@ package ui
 import (
 	"strconv"
 	"strings"
-	"time"
 	"unicode"
 
 	"github.com/eugenioenko/ttt/internal/core/buffer"
@@ -690,542 +689,12 @@ func (e *EditorPaneWidget) pasteText(text string) {
 
 func (e *EditorPaneWidget) HandleEvent(ev tcell.Event) EventResult {
 	if mev, ok := ev.(*tcell.EventMouse); ok {
-		btn := mev.Buttons()
-
-		if newTop, consumed := e.scrollbar.HandleEvent(ev); consumed {
-			if e.Folds != nil && e.Folds.HasCollapsedFolds() {
-				e.Viewport.TopLine = e.Folds.VisibleToBuffer(newTop)
-			} else {
-				e.Viewport.TopLine = newTop
-			}
-			if e.scrollbar.IsDragging() {
-				return EventCaptured
-			}
-			return EventConsumed
-		}
-		if e.scrollbar.IsDragging() {
-			return EventCaptured
-		}
-		if newLeft, consumed := e.hscrollbar.HandleEvent(ev); consumed {
-			e.Viewport.LeftCol = newLeft
-			e.clampLeftCol()
-			if e.hscrollbar.IsDragging() {
-				return EventCaptured
-			}
-			return EventConsumed
-		}
-		if e.hscrollbar.IsDragging() {
-			return EventCaptured
-		}
-
-		mod := mev.Modifiers()
-		if btn&tcell.WheelUp != 0 {
-			if mod&tcell.ModShift != 0 {
-				e.Viewport.LeftCol -= 4
-				e.clampLeftCol()
-			} else {
-				e.scrollUp(3)
-			}
-			return EventConsumed
-		}
-		if btn&tcell.WheelDown != 0 {
-			if mod&tcell.ModShift != 0 {
-				e.Viewport.LeftCol += 4
-				e.clampLeftCol()
-			} else {
-				e.scrollDown(3)
-			}
-			return EventConsumed
-		}
-		if btn&tcell.WheelLeft != 0 {
-			e.Viewport.LeftCol -= 4
-			e.clampLeftCol()
-			return EventConsumed
-		}
-		if btn&tcell.WheelRight != 0 {
-			e.Viewport.LeftCol += 4
-			e.clampLeftCol()
-			return EventConsumed
-		}
-
-		r := e.GetRect()
-		mx, my := mev.Position()
-		gutterW := e.GutterWidth()
-		inGutter := gutterW > 0 && mx >= r.X && mx < r.X+gutterW
-
-		if btn == tcell.ButtonNone && !e.mouseDown {
-			prevHover := e.gutterHover
-			prevLine := e.gutterHoverLine
-			if inGutter {
-				screenY := my - r.Y
-				bufLine := e.screenToBufferLine(screenY)
-				e.gutterHover = true
-				e.gutterHoverLine = bufLine
-			} else {
-				e.gutterHover = false
-			}
-			if e.gutterHover != prevHover || e.gutterHoverLine != prevLine {
-				return EventConsumed
-			}
-		}
-
-		if btn&tcell.Button1 != 0 {
-			if e.Undo != nil {
-				e.Undo.BreakGroup()
-			}
-			line, col := e.mouseToPos(r, mx, my)
-
-			isAlt := mev.Modifiers()&tcell.ModAlt != 0
-			if isAlt && !e.mouseDown {
-				e.ensureMulti()
-				e.syncToMulti()
-				e.Multi.Add(line, col)
-				e.syncFromMulti()
-				e.scrollViewport()
-				return EventCaptured
-			}
-
-			if !e.mouseDown {
-				e.mouseDown = true
-				e.mouseDownX = mx
-				e.mouseDownY = my
-
-				if e.isMultiActive() {
-					e.collapseMulti()
-				}
-
-				now := time.Now().UnixMilli()
-				if now-e.lastClickTime < DoubleClickMs && line == e.lastClickLine && col == e.lastClickCol {
-					e.clickCount++
-				} else {
-					e.clickCount = 1
-				}
-				e.lastClickTime = now
-				e.lastClickLine = line
-				e.lastClickCol = col
-
-				switch e.clickCount {
-				case 2:
-					e.selectWord(line, col)
-				case 3:
-					e.selectLine(line)
-					e.clickCount = 0
-				default:
-					if e.Selection != nil {
-						e.Selection.Clear()
-						e.Selection.Start(line, col)
-					}
-					e.Cursor.Line = line
-					e.Cursor.Col = col
-				}
-			} else {
-				e.Cursor.Line = line
-				e.Cursor.Col = col
-			}
-			e.scrollViewport()
-			return EventCaptured
-		}
-		if btn == tcell.ButtonNone && e.mouseDown {
-			e.mouseDown = false
-			if mx == e.mouseDownX && my == e.mouseDownY && inGutter {
-				bufLine := e.screenToBufferLine(my - r.Y)
-				if e.Folds != nil && e.Folds.FoldAt(bufLine) != nil {
-					e.Folds.Toggle(bufLine)
-					return EventConsumed
-				}
-			}
-			if e.Selection != nil && e.Selection.Active {
-				if e.Selection.Anchor.Line == e.Cursor.Line && e.Selection.Anchor.Col == e.Cursor.Col {
-					e.Selection.Clear()
-				}
-			}
-		}
-		return EventIgnored
+		return e.handleMouse(mev)
 	}
-
-	kev, ok := ev.(*tcell.EventKey)
-	if !ok {
-		return EventIgnored
+	if kev, ok := ev.(*tcell.EventKey); ok {
+		return e.handleKey(kev)
 	}
-
-	e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line)
-
-	switch kev.Key() {
-	case tcell.KeyRune, tcell.KeyBackspace, tcell.KeyBackspace2, tcell.KeyDelete:
-	default:
-		if e.Undo != nil {
-			e.Undo.BreakGroup()
-		}
-	}
-
-	mods := kev.Modifiers()
-	if mods&tcell.ModAlt != 0 || mods&tcell.ModCtrl != 0 {
-		return EventIgnored
-	}
-
-	shift := mods&tcell.ModShift != 0
-	hasSel := e.Selection != nil && e.Selection.Active
-
-	multi := e.isMultiActive()
-
-	switch kev.Key() {
-	case tcell.KeyUp:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				if cs.Line > 0 {
-					cs.Line--
-					if e.Folds != nil {
-						if r := e.Folds.ContainingFold(cs.Line); r != nil {
-							cs.Line = r.StartLine
-						}
-					}
-					lineLen := len([]rune(e.Buf.Lines[cs.Line]))
-					if cs.Col > lineLen {
-						cs.Col = lineLen
-					}
-				}
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			if e.Cursor.Line > 0 {
-				e.Cursor.Line--
-				e.skipHiddenLineUp()
-				lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-				if e.Cursor.Col > lineLen {
-					e.Cursor.Col = lineLen
-				}
-			}
-		}
-	case tcell.KeyDown:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				if cs.Line < len(e.Buf.Lines)-1 {
-					cs.Line++
-					if e.Folds != nil {
-						if r := e.Folds.ContainingFold(cs.Line); r != nil {
-							cs.Line = e.Buf.ClampLine(r.EndLine + 1)
-						}
-					}
-					lineLen := len([]rune(e.Buf.Lines[cs.Line]))
-					if cs.Col > lineLen {
-						cs.Col = lineLen
-					}
-				}
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			if e.Cursor.Line < len(e.Buf.Lines)-1 {
-				e.Cursor.Line++
-				e.skipHiddenLineDown()
-				lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-				if e.Cursor.Col > lineLen {
-					e.Cursor.Col = lineLen
-				}
-			}
-		}
-	case tcell.KeyLeft:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				if cs.Col > 0 {
-					cs.Col--
-				} else if cs.Line > 0 {
-					cs.Line--
-					cs.Col = len([]rune(e.Buf.Lines[cs.Line]))
-				}
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			if e.Cursor.Col > 0 {
-				e.Cursor.Col--
-			} else if e.Cursor.Line > 0 {
-				e.Cursor.Line--
-				e.Cursor.Col = len([]rune(e.Buf.Lines[e.Cursor.Line]))
-			}
-		}
-	case tcell.KeyRight:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				lineLen := len([]rune(e.Buf.Lines[cs.Line]))
-				if cs.Col < lineLen {
-					cs.Col++
-				} else if cs.Line < len(e.Buf.Lines)-1 {
-					cs.Line++
-					cs.Col = 0
-				}
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-			if e.Cursor.Col < lineLen {
-				e.Cursor.Col++
-			} else if e.Cursor.Line < len(e.Buf.Lines)-1 {
-				e.Cursor.Line++
-				e.Cursor.Col = 0
-			}
-		}
-	case tcell.KeyHome:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				runes := []rune(e.Buf.Lines[cs.Line])
-				firstNonSpace := 0
-				for firstNonSpace < len(runes) && (runes[firstNonSpace] == ' ' || runes[firstNonSpace] == '\t') {
-					firstNonSpace++
-				}
-				if cs.Col == firstNonSpace {
-					cs.Col = 0
-				} else {
-					cs.Col = firstNonSpace
-				}
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			e.SmartHome()
-		}
-	case tcell.KeyEnd:
-		if multi {
-			e.multiMoveAll(func(cs *multicursor.CursorState) {
-				cs.Col = len([]rune(e.Buf.Lines[cs.Line]))
-			})
-		} else {
-			e.startOrExtendSelection(shift)
-			e.Cursor.Col = len([]rune(e.Buf.Lines[e.Cursor.Line]))
-		}
-	case tcell.KeyPgUp:
-		e.startOrExtendSelection(shift)
-		e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line - e.Viewport.Height)
-		e.skipHiddenLineUp()
-		lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-		if e.Cursor.Col > lineLen {
-			e.Cursor.Col = lineLen
-		}
-	case tcell.KeyPgDn:
-		e.startOrExtendSelection(shift)
-		e.Cursor.Line = e.Buf.ClampLine(e.Cursor.Line + e.Viewport.Height)
-		e.skipHiddenLineDown()
-		lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-		if e.Cursor.Col > lineLen {
-			e.Cursor.Col = lineLen
-		}
-	case tcell.KeyEnter:
-		e.expandFoldAtCursor()
-		if multi {
-			e.multiExecEnter()
-		} else {
-			if hasSel {
-				e.deleteSelection()
-			}
-			col := e.Cursor.Col
-			if col < 0 {
-				col = 0
-			}
-			lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-			if col > lineLen {
-				col = lineLen
-			}
-			line := e.Buf.Lines[e.Cursor.Line]
-			e.exec(&undo.SplitLineCommand{Line: e.Cursor.Line, Col: col})
-			e.Cursor.Line++
-			e.Cursor.Col = 0
-			if e.AutoIndent {
-				indent := leadingWhitespace(line)
-				runes := []rune(line)
-				charBefore := ' '
-				if col > 0 && col <= len(runes) {
-					charBefore = runes[col-1]
-				}
-				charAfter := ' '
-				if col < len(runes) {
-					charAfter = runes[col]
-				}
-				extraIndent := indentOpeners[charBefore]
-				newIndent := indent
-				if extraIndent {
-					newIndent += e.indentUnit()
-				}
-				if len(newIndent) > 0 {
-					e.exec(&undo.InsertStringCommand{Line: e.Cursor.Line, Col: 0, Text: newIndent})
-					e.Cursor.Col = len([]rune(newIndent))
-				}
-				if extraIndent && closingBrackets[charAfter] {
-					e.exec(&undo.SplitLineCommand{Line: e.Cursor.Line, Col: e.Cursor.Col})
-					e.exec(&undo.InsertStringCommand{Line: e.Cursor.Line + 1, Col: 0, Text: indent})
-				}
-			}
-		}
-	case tcell.KeyBackspace, tcell.KeyBackspace2:
-		e.expandFoldAtCursor()
-		if multi {
-			e.multiExecBackspace()
-		} else {
-			if hasSel {
-				e.deleteSelection()
-			} else if e.Cursor.Col > 0 {
-				runes := []rune(e.Buf.Lines[e.Cursor.Line])
-				if e.Cursor.Col > len(runes) {
-					e.Cursor.Col = len(runes)
-				}
-				inLeadingWhitespace := true
-				for i := 0; i < e.Cursor.Col && i < len(runes); i++ {
-					if runes[i] != ' ' && runes[i] != '\t' {
-						inLeadingWhitespace = false
-						break
-					}
-				}
-				tabSize := e.resolveTabSize()
-				if inLeadingWhitespace && e.Cursor.Col > 1 && runes[e.Cursor.Col-1] == ' ' {
-					target := ((e.Cursor.Col - 1) / tabSize) * tabSize
-					if target == e.Cursor.Col {
-						target -= tabSize
-					}
-					if target < 0 {
-						target = 0
-					}
-					e.exec(&undo.DeleteSelectionCommand{
-						StartLine: e.Cursor.Line, StartCol: target,
-						EndLine: e.Cursor.Line, EndCol: e.Cursor.Col,
-					})
-					e.Cursor.Col = target
-				} else {
-					e.exec(&undo.DeleteRuneCommand{Line: e.Cursor.Line, Col: e.Cursor.Col - 1})
-					e.Cursor.Col--
-				}
-			} else if e.Cursor.Line > 0 {
-				cmd := &undo.JoinLineCommand{Line: e.Cursor.Line}
-				e.exec(cmd)
-				e.Cursor.Line--
-				e.Cursor.Col = cmd.PrevLen
-			}
-		}
-	case tcell.KeyDelete:
-		e.expandFoldAtCursor()
-		if multi {
-			e.multiExecDelete()
-		} else {
-			if hasSel {
-				e.deleteSelection()
-			} else {
-				lineLen := len([]rune(e.Buf.Lines[e.Cursor.Line]))
-				if e.Cursor.Col < lineLen {
-					e.exec(&undo.DeleteRuneCommand{Line: e.Cursor.Line, Col: e.Cursor.Col})
-				} else if e.Cursor.Line < len(e.Buf.Lines)-1 {
-					e.exec(&undo.JoinLineCommand{Line: e.Cursor.Line + 1})
-				}
-			}
-		}
-	case tcell.KeyRune:
-		if kev.Modifiers() == 0 {
-			e.expandFoldAtCursor()
-			r := kev.Rune()
-			if r != 0 {
-				if multi {
-					e.multiExecRune(r)
-				} else if hasSel {
-					start, _ := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-					e.replaceSelection(&undo.InsertRuneCommand{Line: start.Line, Col: start.Col, Rune: r})
-					e.Cursor.Col = start.Col + 1
-				} else {
-					if e.AutoIndent && closingBrackets[r] {
-						e.dedentForCloser()
-					}
-					e.exec(&undo.InsertRuneCommand{Line: e.Cursor.Line, Col: e.Cursor.Col, Rune: r})
-					e.Cursor.Col++
-				}
-			}
-		} else {
-			return EventIgnored
-		}
-	case tcell.KeyBacktab:
-		tabSize := e.resolveTabSize()
-		if hasSel {
-			start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-			end.Line = e.Buf.ClampLine(end.Line)
-			for line := start.Line; line <= end.Line; line++ {
-				remove := leadingIndentWidth(e.Buf.Lines[line], tabSize)
-				if remove > 0 {
-					e.exec(&undo.DeleteSelectionCommand{
-						StartLine: line, StartCol: 0,
-						EndLine: line, EndCol: remove,
-					})
-				}
-			}
-		} else {
-			remove := leadingIndentWidth(e.Buf.Lines[e.Cursor.Line], tabSize)
-			if remove > 0 {
-				e.exec(&undo.DeleteSelectionCommand{
-					StartLine: e.Cursor.Line, StartCol: 0,
-					EndLine: e.Cursor.Line, EndCol: remove,
-				})
-				e.Cursor.Col -= remove
-				if e.Cursor.Col < 0 {
-					e.Cursor.Col = 0
-				}
-			}
-		}
-	case tcell.KeyTab:
-		indent := e.indentUnit()
-		if hasSel {
-			start, end := e.Selection.Range(e.Cursor.Line, e.Cursor.Col)
-			for line := start.Line; line <= end.Line; line++ {
-				e.exec(&undo.InsertStringCommand{Line: line, Col: 0, Text: indent})
-			}
-		} else {
-			e.exec(&undo.InsertStringCommand{Line: e.Cursor.Line, Col: e.Cursor.Col, Text: indent})
-			e.Cursor.Col += len([]rune(indent))
-		}
-	default:
-		return EventIgnored
-	}
-
-	e.clampCursor()
-	e.scrollViewport()
-	return EventConsumed
-}
-
-func (e *EditorPaneWidget) mouseToPos(r Rect, mx, my int) (line, col int) {
-	if len(e.Buf.Lines) == 0 {
-		return 0, 0
-	}
-	gutterW := e.GutterWidth()
-	screenY := my - r.Y
-
-	if e.WordWrap && e.wrapMap != nil && screenY >= 0 && screenY < len(e.wrapMap) {
-		entry := e.wrapMap[screenY]
-		line = entry.bufLine
-		if line >= len(e.Buf.Lines) {
-			line = len(e.Buf.Lines) - 1
-		}
-		segVisCol := mx - r.X - gutterW
-		if segVisCol < 0 {
-			segVisCol = 0
-		}
-		segLeftCol := bufColToVisualCol(e.Buf.Lines[line], entry.startCol, e.resolveTabSize())
-		col = visualColToBufCol(e.Buf.Lines[line], segLeftCol+segVisCol, e.resolveTabSize())
-	} else {
-		line = e.screenToBufferLine(screenY)
-		visCol := mx - r.X - gutterW + e.Viewport.LeftCol
-		if visCol < 0 {
-			visCol = 0
-		}
-		if line < 0 {
-			line = 0
-		}
-		if line >= len(e.Buf.Lines) {
-			line = len(e.Buf.Lines) - 1
-		}
-		col = visualColToBufCol(e.Buf.Lines[line], visCol, e.resolveTabSize())
-	}
-
-	if line < 0 {
-		line = 0
-	}
-	if line >= len(e.Buf.Lines) {
-		line = len(e.Buf.Lines) - 1
-	}
-	lineLen := len([]rune(e.Buf.Lines[line]))
-	if col > lineLen {
-		col = lineLen
-	}
-	return
+	return EventIgnored
 }
 
 func (e *EditorPaneWidget) selectWord(line, col int) {
